Update Holacracy-Constitution.md
I'm a bit confuse with the “people” word, as we basically have Partner/Partners and external parties in the Constitution, so is the “people” (as it's not defined in the Constitution, just like “external parties” in fact) term the “union” of the two? In which case, I'd like to clarify, or if it concerns only Partners, then fine, but it's a bit fuzzy for me.